chore: Update eslint configuration and header navigation links
